+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/handlerMapping/HandlerMapping.java
@@ -1,4 +1,4 @@
-package webmvc.org.springframework.web.servlet.mvc.tobe;
+package webmvc.org.springframework.web.servlet.mvc.handlerMapping;
 
 import jakarta.servlet.http.HttpServletRequest;
