+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/view/JspView.java
@@ -6,7 +6,6 @@ import jakarta.servlet.http.HttpServletResponse;
 import java.util.Map;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import webmvc.org.springframework.web.servlet.View;
 
 public class JspView implements View {
